Refactor Invitation and Event entities for improved relationships and data handling

+++ apifilrouge/src/main/java/com/projetfilrougeapi/apifilrouge/endpoint_api/event/EventRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @RepositoryRestResource(collectionResourceRel = "events", path = "events")
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
-    Optional<Event> findById(Integer id);
+public interface EventRepository extends JpaRepository<Event, Long> {
+    Optional<Event> findById(Long id);
 }
